feat: Add support for data buttons in map controls
This commit adds support for data buttons in the map controls component. The code changes include:
- Importing the necessary types for data buttons
- Adding a new ng-container to handle data buttons
- Iterating over the data buttons and rendering them using wm-map-button-control component
- Emitting an event when a data button is clicked, along with its type and toggle status
- Initializing the default state of data buttons based on their "default" property

These changes allow users to interact with data buttons in the map controls component and toggle their visibility.

feat: add toggleData event to map component

This commit adds a new event called `toggleData` to the map component. This event is triggered when the user interacts with the map and toggles the display of layers or points of interest. The event emits an object containing information about the type of data (layers or pois) and whether it should be toggled on or off.

chore: Refactor layer.directive.ts

Simplified the code by removing unnecessary conditionals and improving readability.

feat(types): add optional properties to ICONTROLSBUTTON interface

This commit adds two optional properties, 'icon' and 'id', to the ICONTROLSBUTTON interface in the model.ts file. Additionally, it introduces the 'default' property as an optional property for the 'label' field.

chore: Refactor map component to handle changes more efficiently

- Added private methods to handle specific property changes
- Consolidated code in ngOnChanges method for better readability and maintainability
- Improved the handling of filters, reset state, and wmMapPadding changes
- Created a new private method to update the map's state and trigger necessary actions

chore: Refactor controls.map