feat(map.component): add fullscreen control option to map
Added a new input property `wmMapFullscreen` to the `WmMapComponent`. This allows for enabling or disabling a fullscreen control on the map. The default value is false, meaning the fullscreen control will not be displayed unless explicitly enabled. Also refactored how controls are added to the map for better readability and maintainability.