feat: add Google Maps navigation to POIs
Introduced a new feature in the Points of Interest (POIs) directive that allows users to navigate to a selected POI using Google Maps. This is achieved by adding a new button in the tooltip popup which, when clicked, opens Google Maps in a new tab with directions to the selected POI. Also added 'coordsToLonLat' utility function import for coordinate conversion.