feat: Add support for creating POI markers from FeatureCollection oc: 2534
This commit adds a new input property `wmMapPoisPois` to the `WmMapTrackRelatedPoisDirective` directive. This property allows users to pass a `FeatureCollection` object containing POI data, which will be used to create POI markers on the map.

The `_createPoiMarker()` method has been updated to handle the creation of POI markers using the provided `FeatureCollection`. The method reads features from the collection and creates corresponding marker objects with appropriate styles.

Additionally, a private method `_getIcnFromTaxonomies()` has been added to retrieve the ICN (Icon Identifier) from taxonomy identifiers associated with each POI. This ICN is used to determine the icon source for the marker.

These changes enhance the functionality of the directive by allowing users to dynamically create and display POI markers on the map based on provided data.