feat(layer.directive): add zoom on click functionality
This commit adds the functionality to zoom in when a user clicks on the map. If the current zoom level is below a certain threshold, it will zoom in to a default zoom level. The code also includes some refactoring and bug fixes related to handling feature clicks.

BREAKING CHANGE: The constant `MAP_ZOOM_ON_CLICK_TRESHOLD` has been added to define the threshold for zooming on click. The constant `DEF_ZOOM_ON_CLICK` has been added to define the default zoom level when clicking on the map.

Closes #123