feat: Add fitView caller parameter to WmMapBaseDirective
This commit adds a new optional parameter "caller" to the fitView method in the WmMapBaseDirective. The "caller" parameter is used to determine whether or not to execute the fitView function based on certain conditions. This change allows for more control over when the fitView function is called.

chore: Refactor WmMapLayerDirective and WmMapTrackDirective

This commit refactors the code in the WmMapLayerDirective and WmMapTrackDirective. It updates how the fitView function is called in both directives, taking into account the new "caller" parameter introduced in the previous commit. This change ensures that the fitView function is only executed under specific conditions, improving performance and functionality.

Note: These commits are part of ongoing development and may require further testing and refinement.