chore(hit-map): add tile layer above existing layer
This commit adds a new tile layer above the existing layer in the hit map directive. The tile layer uses the same tile grid and projection as the existing layer. It is added to the map with an increased zIndex to position it above the existing layer.

BREAKING CHANGE: None