feat(hit-map): add new TileLayer with updated properties
This commit adds a new TileLayer to the hit map directive. The new TileLayer uses a different URL for the tile source and has an updated projection. It is also set to be visible with an opacity of 1 and a zIndex of 1. This change improves the functionality and appearance of the hit map directive.