refactor(draw-ugc-poi-directive): ♻️ improve event handling and input management
Revised the event handling mechanism by adding a private _ugcPoidrawn property to manage the state of the UGC POI being drawn. Modified the event listener setup to dynamically attach and detach the 'click' handler based on the directive's enabled state. Replaced the use of ngOnChanges with a setter for the wmMapDrawUgcPoiPoi input to immediately reflect any changes in the UGC POI data. This enhances performance and ensures that the map component is only responding to the click events when necessary, reducing unnecessary operations. Additionally, updated the event emitter to use wmMapDrawUgcPoiEvt for consistency and removed redundant code for setting layer visibility within the _init method.