chore(directives): ✨ add UGC drawing directive oc:5655 (#36)
* chore(directives): ✨ add UGC drawing directive

Introduce `WmMapDrawUgcDirective` to enable drawing User Generated Content (UGC) such as Points of Interest (POI) and tracks on the map. The directive manages the display of UGC layers, allowing interaction through clicks to place or update UGC POIs.

- Added `WmMapDrawUgcDirective` to handle UGC drawing and interaction.
- Updated `map-core.module.ts` to include the new directive.
- Defined `UGC_POI_DRAWN_ZINDEX` in `zIndex.ts` for proper layer stacking.

The directive supports toggling visibility and selecting POI icons, and will be extended to handle UGC track interactions in the future.

refactor(directives): 🔄 rename and simplify UGC directive

Renamed `draw-ugc.directive.ts` to `draw-ugc-poi.directive.ts` and refactored the directive to focus solely on handling UGC POI (Point of Interest) features. Removed handling of UGC track features to simplify the directive.

Changes include:
- Updated the selector and class name to reflect the focus on POIs.
- Removed the UGC track handling logic and related inputs.
- Simplified methods and property names for clarity.
- Updated module imports to reflect the new directive name.

refactor(directives): ♻️ rename and optimize POI handling logic

This commit refactors the POI handling in the `WmMapDrawUgcPoiDirective` and `WmUgcPoisDirective` by renaming variables and methods for better clarity and consistency. The `_ugcPoi` behavior subject and the `ugcPoi` event emitter are renamed to better reflect their purpose, and the logic for handling map changes and POI drawing is streamlined.

Additionally, the `wmMapUgcPoi` input in `WmUgcPoisDirective` is updated to accept `WmFeature<Point>` instead of a number or string, aligning with the refactored logic.

- Renamed `_ugcPoiIcon` to `_drawUgcPoiIcon` for better clarity.
- Replaced `_setPoi` method with direct use of `_selectIcon`.
- Updated `UGC_POI_DRAWN_ZINDEX` for better layering control.

* refactor(draw-ugc-poi-directive): ♻️ improve event handling and input management

Revised the event handling mechanism by adding a private _ugcPoidrawn property to manage the state of the UGC POI being drawn. Modified the event listener setup to dynamically attach and detach the 'click' handler based on the directive's enabled state. Replaced the use of ngOnChanges with a setter for the wmMapDrawUgcPoiPoi input to immediately reflect any changes in the UGC POI data. This enhances performance and ensures that the map component is only responding to the click events when necessary, reducing unnecessary operations. Additionally, updated the event emitter to use wmMapDrawUgcPoiEvt for consistency and removed redundant code for setting layer visibility within the _init method.