chore(utils): ✨ add distance calculation and optimize arrow filtering oc:5800 (#46)
* chore(utils): ✨ add distance calculation and optimize arrow filtering

Added a new utility function `calculateDistance` to compute the Euclidean distance between two points. This function is now used in `styles.ts` to optimize the filtering of arrow styles based on their proximity to the current track.

- Introduced `calculateDistance` in `geometry.ts` for computing distances between coordinates.
- Updated `styleFn` in `styles.ts` to filter arrow styles more efficiently, using the distance calculation to remove points too close to `currentTrack`.
- Integrated coordinate transformations to ensure accurate distance measurements in the map's coordinate system.

These enhancements improve performance by minimizing unnecessary rendering of arrow styles that are too close to the current track, particularly when the track is displayed without routing.

* chore(geometry): ✨ add feature conversion to Web Mercator

Add a utility function `convertFeatureToWebMercator` to transform features from EPSG:4326 (WGS84) to EPSG:3857 (Web Mercator) without mutating the original feature. This function is utilized in `WmMapLayerDirective` to convert `currentTrack` before processing.

refactor(styles): ♻️ remove redundant coordinate transformations

Remove unnecessary transformations between EPSG:4326 and EPSG:3857 in the `styleFn`. The coordinates are already in EPSG:3857, so we directly use them to calculate the closest point and distance, simplifying the logic and improving performance.

* refactor(geometry): 🔄 rename conversion and distance functions

Renamed `convertFeatureToWebMercator` to `convertFeatureToEpsg3857` for clarity and consistency with EPSG code. Similarly, renamed `calculateDistance` to `calculatePointsDistance` for better understanding of its functionality.

refactor(styles): ♻️ extract arrow style filtering logic

Moved the arrow style filtering logic to a dedicated function `getFilteredDirectionArrows` for better readability and modularity. This function generates and filters direction arrow styles for a track by removing those that are too close to the current track.