chore(directives): ✨ add wmMapDisableFitView input and improve geojson layer handling oc: 5239 (#47)
* chore(directives): ✨ add wmMapDisableFitView input and improve geojson layer handling

Added a new input property `wmMapDisableFitView` to the `WmMapBaseDirective` to allow disabling the automatic fitting of the map view. This provides more control over the map behavior.

Additionally, improved the `WmMapGeojsonDirective` to handle cases where the feature collection is null by removing the existing layer and resetting the initialization state. This prevents stale layers from persisting and ensures the directive's state is accurately reflected.

* refactor(directives): ♻️ simplify feature collection layer reset logic

Simplified the logic for resetting the feature collection layer in the `WmMapGeojsonDirective`. Instead of checking if the layer is not null and then removing it, the new approach directly clears the source if it exists, which reduces redundancy and improves readability.